Move Activator to shared partition (#22894)
