+++ src/System.Private.CoreLib/src/System/RtType.cs
@@ -4328,7 +4328,7 @@ namespace System
         }
 
         internal object CreateInstanceImpl(
-            BindingFlags bindingAttr, Binder binder, object[] args, CultureInfo culture, object[] activationAttributes)
+            BindingFlags bindingAttr, Binder binder, object[] args, CultureInfo culture)
         {
             CreateInstanceCheckThis();
