+++ src/System.Private.CoreLib/src/System/__ComObject.cs
@@ -147,7 +147,7 @@ namespace System
         private object CreateEventProvider(RuntimeType t)
         {
             // Create the event provider for the specified type.
-            object EvProvider = Activator.CreateInstance(t, Activator.ConstructorDefault | BindingFlags.NonPublic, null, new object[] { this }, null);
+            object EvProvider = Activator.CreateInstance(t, BindingFlags.Instance | BindingFlags.Public | BindingFlags.NonPublic | BindingFlags.CreateInstance, null, new object[] { this }, null);
 
             // Attempt to cache the wrapper on the object.
             if (!SetData(t, EvProvider))
